lxd: Move `MergeSSHKeyCloudConfig` to new `cloudinit` package
